+++ Kernel/Modules/AgentTicketCompose.pm
@@ -936,12 +936,7 @@ sub Run {
             ATTACHMENT:
             for my $Attachment (@AttachmentData) {
                 my $ContentID = $Attachment->{ContentID};
-                if (
-                    $ContentID
-                    && ( $Attachment->{ContentType} =~ /image/i )
-                    && ( $Attachment->{Disposition} eq 'inline' )
-                    )
-                {
+                if ( $ContentID && ( $Attachment->{ContentType} =~ /image/i ) ) {
                     my $ContentIDHTMLQuote = $LayoutObject->Ascii2Html(
                         Text => $ContentID,
                     );
@@ -950,9 +945,15 @@ sub Run {
                     my $ContentIDLinkEncode = $LayoutObject->LinkEncode($ContentID);
                     $GetParam{Body} =~ s/(ContentID=)$ContentIDLinkEncode/$1$ContentID/g;
 
-                    # ignore attachment if not linked in body
-                    next ATTACHMENT
-                        if $GetParam{Body} !~ /(\Q$ContentIDHTMLQuote\E|\Q$ContentID\E)/i;
+                    # IF the image is referenced in the body set it as inline.
+                    if ( $GetParam{Body} =~ /(\Q$ContentIDHTMLQuote\E|\Q$ContentID\E)/i ) {
+                        $Attachment->{Disposition} = 'inline';
+                    }
+                    elsif ( $Attachment->{Disposition} eq 'inline' ) {
+
+                        # Ignore attachment if not linked in body.
+                        next ATTACHMENT;
+                    }
                 }
 
                 # remember inline images and normal attachments